Split out isPlan
